+++ dashboard/src/components/partials/EntityDetailsContainer/EntityDetailsContainer.js
@@ -1,9 +1,13 @@
 import React from "react";
 import PropTypes from "prop-types";
 import gql from "graphql-tag";
-import Grid from "@material-ui/core/Grid";
+
+import CollapsingMenu from "/components/CollapsingMenu";
 import Content from "/components/Content";
-import ButtonSet from "/components/ButtonSet";
+import Grid from "@material-ui/core/Grid";
+import DeleteIcon from "@material-ui/icons/Delete";
+import LiveIcon from "/icons/Live";
+
 import RelatedEntitiesCard from "/components/partials/RelatedEntitiesCard";
 import EntityDetailsEvents from "./EntityDetailsEvents";
 import EntityDetailsInformation from "./EntityDetailsInformation";
@@ -12,6 +16,11 @@ import DeleteAction from "./EntityDetailsDeleteAction";
 class EntityDetailsContainer extends React.PureComponent {
   static propTypes = {
     entity: PropTypes.object.isRequired,
+    poller: PropTypes.shape({
+      running: PropTypes.bool,
+      start: PropTypes.func,
+      stop: PropTypes.func,
+    }).isRequired,
   };
 
   static fragments = {
@@ -35,14 +44,29 @@ class EntityDetailsContainer extends React.PureComponent {
   };
 
   render() {
-    const { entity } = this.props;
+    const { entity, poller } = this.props;
+
     return (
       <React.Fragment>
         <Content bottomMargin>
           <div style={{ flexGrow: 1 }} />
-          <ButtonSet>
-            <DeleteAction entity={entity} />
-          </ButtonSet>
+          <CollapsingMenu>
+            <DeleteAction entity={entity}>
+              {del => (
+                <CollapsingMenu.Button
+                  title="Delete"
+                  icon={<DeleteIcon />}
+                  onClick={() => del()}
+                />
+              )}
+            </DeleteAction>
+            <CollapsingMenu.Button
+              pinned
+              title="LIVE"
+              icon={<LiveIcon active={poller.running} />}
+              onClick={() => (poller.running ? poller.stop() : poller.start())}
+            />
+          </CollapsingMenu>
         </Content>
         <Content>
           <Grid container spacing={16}>