+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 - Added sensuctl cluster member-list command.
 - Added Sensu edition detection in sensuctl.
 - Added sensuctl cluster member-add command.
+- Added API client support for enterprise license management.
 
 ### Changed
 - The Backend struct has been refactored to allow easier customization in